Revert "Expose internal functionality for lua-stormlib"
This reverts commit fba8fca3ac2691906c081c7850f44d4f50203b41.

A bit of tunnel vision kicked in regarding the reverted commit.  An
attempt was made to adapt a method that would have allowed StormLib to
update file references internally when increasing an archive's file
limit.  And thus the previous commit.

However, it was realized that the information necessary to update the
references could be gleaned via the existing public API.  That is, by
searching for the files by name, as the resulting structure contains
such data (i.e. both the hash and block table indices).  Thus life is
now simple again.